Added Stations
In-progress work. Stations are accessed by pressing "E"; it's a little fiddly right now but works most of the time. Oar station makes the left and right inputs for movement apply to the raft and be much stronger, and the harpoon station extends the reach of the harpoon drastically and speeds it up.